Improved, modified and fixed

+++ usci/modules/eav/modules/persistance/src/main/java/kz/bsbnb/usci/eav/persistance/dao/impl/UserDaoImpl.java
@@ -3,8 +3,10 @@ package kz.bsbnb.usci.eav.persistance.dao.impl;
 import kz.bsbnb.usci.cr.model.Creditor;
 import kz.bsbnb.usci.cr.model.PortalUser;
 import kz.bsbnb.usci.cr.model.SubjectType;
+import kz.bsbnb.usci.eav.model.base.IBaseEntity;
 import kz.bsbnb.usci.eav.model.base.impl.BaseEntity;
 import kz.bsbnb.usci.eav.model.base.impl.BaseValue;
+import kz.bsbnb.usci.eav.persistance.dao.IBaseEntityLoadDao;
 import kz.bsbnb.usci.eav.persistance.dao.IBaseEntityProcessorDao;
 import kz.bsbnb.usci.eav.persistance.dao.IUserDao;
 import kz.bsbnb.usci.eav.persistance.db.JDBCSupport;
@@ -20,18 +22,13 @@ import org.springframework.stereotype.Repository;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
-import java.sql.*;
+import java.sql.Timestamp;
 import java.util.*;
-import java.util.Date;
 
 import static kz.bsbnb.eav.persistance.generated.Tables.*;
 
-/**
- *
- */
 @Repository
-public class UserDaoImpl extends JDBCSupport implements IUserDao
-{
+public class UserDaoImpl extends JDBCSupport implements IUserDao {
     private final Logger logger = LoggerFactory.getLogger(UserDaoImpl.class);
 
     @SuppressWarnings("SpringJavaAutowiringInspection")
@@ -41,132 +38,94 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
     @Autowired
     IBaseEntityProcessorDao baseEntityProcessorDao;
 
+    @Autowired
+    IBaseEntityLoadDao baseEntityLoadDao;
+
     @Override
-    public boolean hasPortalUserCreditor(long userId, long creditorId)
-    {
+    public boolean hasPortalUserCreditor(long userId, long creditorId) {
         SelectForUpdateStep select;
 
-
-
-        select = context.select(
-                EAV_A_CREDITOR_USER.ID
-        ).from(EAV_A_CREDITOR_USER).
+        select = context.select(EAV_A_CREDITOR_USER.ID).from(EAV_A_CREDITOR_USER).
                 where(EAV_A_CREDITOR_USER.USER_ID.eq(userId)).and(EAV_A_CREDITOR_USER.CREDITOR_ID.eq(creditorId));
 
-
-        logger.debug(select.toString());
         List<Map<String, Object>> rows = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
 
-
-        if (rows.size() > 0)
-            return true;
-
-        return false;
+        return rows.size() > 0;
     }
 
     @Override
-    public void setPortalUserCreditors(long userId, long creditorId)
-    {
-        try
-        {
-            InsertOnDuplicateStep insert = context.insertInto(
-                    EAV_A_CREDITOR_USER,
-                    EAV_A_CREDITOR_USER.USER_ID,
-                    EAV_A_CREDITOR_USER.CREDITOR_ID
-            ).values(userId, creditorId);
+    public void setPortalUserCreditors(long userId, long creditorId) {
+        InsertOnDuplicateStep insert = context.insertInto(
+                EAV_A_CREDITOR_USER,
+                EAV_A_CREDITOR_USER.USER_ID,
+                EAV_A_CREDITOR_USER.CREDITOR_ID
+        ).values(userId, creditorId);
 
-            logger.debug(insert.toString());
-            insertWithId(insert.getSQL(), insert.getBindValues().toArray());
-        }
-        catch (DuplicateKeyException e)
-        {
-            logger.error("Duplicate ids: " + userId + " " + creditorId);
-            throw new IllegalArgumentException("Duplicate ids: " + userId + " " + creditorId);
-        }
+        insertWithId(insert.getSQL(), insert.getBindValues().toArray());
     }
 
     @Override
-    public void unsetPortalUserCreditors(long userId, long creditorId)
-    {
-        DeleteConditionStep deleteFilter =
-                context.delete(EAV_A_CREDITOR_USER).
-                        where(EAV_A_CREDITOR_USER.USER_ID.eq(userId)).and(EAV_A_CREDITOR_USER.CREDITOR_ID.eq(creditorId));
-
-        long t = 0;
-        if(sqlStats != null)
-        {
-            t = System.nanoTime();
-        }
+    public void unsetPortalUserCreditors(long userId, long creditorId) {
+        DeleteConditionStep deleteFilter = context.delete(EAV_A_CREDITOR_USER).
+                where(EAV_A_CREDITOR_USER.USER_ID.eq(userId)).
+                and(EAV_A_CREDITOR_USER.CREDITOR_ID.eq(creditorId));
 
         jdbcTemplate.update(deleteFilter.getSQL(), deleteFilter.getBindValues().toArray());
-
-        if(sqlStats != null)
-        {
-            sqlStats.put(deleteFilter.getSQL(), (System.nanoTime() - t) / 1000000);
-        }
     }
 
     @Override
-    public List<Creditor> getPortalUserCreditorList(long userId)
-    {
-        ArrayList<Creditor> creditors = new ArrayList<Creditor>();
+    public List<Creditor> getPortalUserCreditorList(long userId) {
+        ArrayList<Creditor> creditors = new ArrayList<>();
 
         SelectForUpdateStep select;
 
-
-
-        select = context.select(
-                EAV_A_CREDITOR_USER.CREDITOR_ID
-        ).from(EAV_A_CREDITOR_USER).
+        select = context.select(EAV_A_CREDITOR_USER.CREDITOR_ID).from(EAV_A_CREDITOR_USER).
                 where(EAV_A_CREDITOR_USER.USER_ID.eq(userId));
 
-
-        logger.debug(select.toString());
         List<Map<String, Object>> rows = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
-
 
         if (rows.size() < 1)
             return creditors;
 
-        for (Map<String, Object> row : rows){
-            Long id = ((BigDecimal)row.get(EAV_A_CREDITOR_USER.CREDITOR_ID.getName())).longValue();
+        for (Map<String, Object> row : rows) {
+            Long id = ((BigDecimal) row.get(EAV_A_CREDITOR_USER.CREDITOR_ID.getName())).longValue();
 
-            BaseEntity entity = (BaseEntity) baseEntityProcessorDao.load(id);
+            IBaseEntity entity = baseEntityLoadDao.load(id);
 
             Creditor creditor = new Creditor();
 
             creditor.setId(entity.getId());
-            BaseValue value = (BaseValue)entity.getBaseValue("name");
+            BaseValue value = (BaseValue) entity.getBaseValue("name");
             if (value != null)
-                creditor.setName((String)value.getValue());
+                creditor.setName((String) value.getValue());
             else
                 creditor.setName("none");
 
-            value = (BaseValue)entity.getBaseValue("short_name");
+            value = (BaseValue) entity.getBaseValue("short_name");
             if (value != null)
-                creditor.setShortName((String)value.getValue());
+                creditor.setShortName((String) value.getValue());
             else
                 creditor.setShortName("none");
 
-            value = (BaseValue)entity.getBaseValue("code");
+            value = (BaseValue) entity.getBaseValue("code");
             if (value != null)
-                creditor.setCode((String)value.getValue());
+                creditor.setCode((String) value.getValue());
             else
                 creditor.setCode("none");
 
             SubjectType st = new SubjectType();
-            BaseValue val = (BaseValue)entity.getBaseValue("subject_type");
-            BaseEntity stEntity = val == null ? null : (BaseEntity)val.getValue();
+            BaseValue val = (BaseValue) entity.getBaseValue("subject_type");
+            BaseEntity stEntity = val == null ? null : (BaseEntity) val.getValue();
             if (stEntity != null) {
                 st.setId(stEntity.getId());
 
-                for(String s : stEntity.getAttributes()) {
+                for (String s : stEntity.getAttributes()) {
                     Object obj = stEntity.getBaseValue(s).getValue();
                     if (obj == null) {
                         continue;
                     }
                     if (s.equals("code")) {
-                        st.setCode((String)obj);
+                        st.setCode((String) obj);
                     } else if (s.equals("name_ru")) {
                         st.setNameRu((String) obj);
                     } else if (s.equals("name_kz")) {
@@ -186,8 +145,6 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
             }
 
             creditor.setSubjectType(st);
-
-
             creditors.add(creditor);
         }
 
@@ -195,14 +152,11 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
     }
 
     @Override
-    public void synchronize(List<PortalUser> users)
-    {
-        HashMap<Long, PortalUser> usersFromDB = new HashMap<Long, PortalUser>();
-        HashMap<Long, PortalUser> usersFromPortal = new HashMap<Long, PortalUser>();
+    public void synchronize(List<PortalUser> users) {
+        HashMap<Long, PortalUser> usersFromDB = new HashMap<>();
+        HashMap<Long, PortalUser> usersFromPortal = new HashMap<>();
 
         SelectForUpdateStep select;
-
-
 
         select = context.select(
                 EAV_A_USER.ID,
@@ -215,29 +169,24 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
                 EAV_A_USER.MODIFIED_DATE
         ).from(EAV_A_USER);
 
-
-        logger.debug(select.toString());
         List<Map<String, Object>> rows = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
 
-
-        for (Map<String, Object> row : rows){
+        for (Map<String, Object> row : rows) {
             PortalUser user = new PortalUser();
 
             user.setId(BigInteger.valueOf(Long.parseLong(row.get(EAV_A_USER.ID.getName()).toString())));
             user.setUserId(Long.parseLong(row.get(EAV_A_USER.USER_ID.getName()).toString()));
-            user.setFirstName((String)row.get(EAV_A_USER.FIRST_NAME.getName()));
+            user.setFirstName((String) row.get(EAV_A_USER.FIRST_NAME.getName()));
             user.setLastName((String) row.get(EAV_A_USER.LAST_NAME.getName()));
             user.setMiddleName((String) row.get(EAV_A_USER.MIDDLE_NAME.getName()));
-            user.setScreenName((String)row.get(EAV_A_USER.SCREEN_NAME.getName()));
-            user.setEmailAddress((String)row.get(EAV_A_USER.EMAIL.getName()));
-            user.setModifiedDate(DataUtils.convert((Timestamp)row.get(EAV_A_USER.MODIFIED_DATE.getName())));
+            user.setScreenName((String) row.get(EAV_A_USER.SCREEN_NAME.getName()));
+            user.setEmailAddress((String) row.get(EAV_A_USER.EMAIL.getName()));
+            user.setModifiedDate(DataUtils.convert((Timestamp) row.get(EAV_A_USER.MODIFIED_DATE.getName())));
 
             usersFromDB.put(user.getUserId(), user);
         }
 
-        for (PortalUser user : users) {
-            usersFromPortal.put(user.getUserId(), user);
-        }
+        for (PortalUser user : users) usersFromPortal.put(user.getUserId(), user);
 
         Set<Long> toDelete = SetUtils.difference(usersFromDB.keySet(), usersFromPortal.keySet());
         Set<Long> toAdd = SetUtils.difference(usersFromPortal.keySet(), usersFromDB.keySet());
@@ -248,69 +197,36 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
                     context.delete(EAV_A_CREDITOR_USER).
                             where(EAV_A_CREDITOR_USER.USER_ID.eq(id));
 
-            long t = 0;
-            if(sqlStats != null)
-            {
-                t = System.nanoTime();
-            }
-
             jdbcTemplate.update(deleteFilter.getSQL(), deleteFilter.getBindValues().toArray());
 
-            if(sqlStats != null)
-            {
-                sqlStats.put(deleteFilter.getSQL(), (System.nanoTime() - t) / 1000000);
-            }
-
-            deleteFilter =
-                    context.delete(EAV_A_USER).
-                            where(EAV_A_USER.USER_ID.eq(id));
-
-            t = 0;
-            if(sqlStats != null)
-            {
-                t = System.nanoTime();
-            }
+            deleteFilter = context.delete(EAV_A_USER).where(EAV_A_USER.USER_ID.eq(id));
 
             jdbcTemplate.update(deleteFilter.getSQL(), deleteFilter.getBindValues().toArray());
-
-            if(sqlStats != null)
-            {
-                sqlStats.put(deleteFilter.getSQL(), (System.nanoTime() - t) / 1000000);
-            }
         }
 
         for (Long id : toAdd) {
-            try
-            {
-                PortalUser pu = usersFromPortal.get(id);
+            PortalUser pu = usersFromPortal.get(id);
 
-                InsertOnDuplicateStep insert = context.insertInto(
-                        EAV_A_USER,
-                        EAV_A_USER.USER_ID,
-                        EAV_A_USER.FIRST_NAME,
-                        EAV_A_USER.LAST_NAME,
-                        EAV_A_USER.MIDDLE_NAME,
-                        EAV_A_USER.SCREEN_NAME,
-                        EAV_A_USER.EMAIL,
-                        EAV_A_USER.MODIFIED_DATE
-                ).values(
-                        pu.getUserId(),
-                        pu.getFirstName(),
-                        pu.getLastName(),
-                        pu.getMiddleName(),
-                        pu.getScreenName(),
-                        pu.getEmailAddress(),
-                        DataUtils.convert(pu.getModifiedDate())
-                );
+            InsertOnDuplicateStep insert = context.insertInto(
+                    EAV_A_USER,
+                    EAV_A_USER.USER_ID,
+                    EAV_A_USER.FIRST_NAME,
+                    EAV_A_USER.LAST_NAME,
+                    EAV_A_USER.MIDDLE_NAME,
+                    EAV_A_USER.SCREEN_NAME,
+                    EAV_A_USER.EMAIL,
+                    EAV_A_USER.MODIFIED_DATE
+            ).values(
+                    pu.getUserId(),
+                    pu.getFirstName(),
+                    pu.getLastName(),
+                    pu.getMiddleName(),
+                    pu.getScreenName(),
+                    pu.getEmailAddress(),
+                    DataUtils.convert(pu.getModifiedDate())
+            );
 
-                logger.debug(insert.toString());
-                insertWithId(insert.getSQL(), insert.getBindValues().toArray());
-            }
-            catch (DuplicateKeyException e)
-            {
-                logger.error("Duplicate id: " + id);
-                throw new IllegalArgumentException("Duplicate id: " + id);
-            }
+            insertWithId(insert.getSQL(), insert.getBindValues().toArray());
         }
 
         for (Long id : toUpdate) {
@@ -328,7 +244,6 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
                     .where(EAV_A_USER.USER_ID.equal(id))
                     .and(EAV_A_USER.MODIFIED_DATE.lessThan(DataUtils.convert(pu.getModifiedDate())));
 
-            logger.debug(update.toString());
             updateWithStats(update.getSQL(), update.getBindValues().toArray());
         }
     }
@@ -339,44 +254,39 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
                 .from(EAV_A_USER_CLASS)
                 .where(EAV_A_USER_CLASS.USER_ID.eq(portalUserId));
 
-        List<String> list = new ArrayList<String>();
+        List<String> list = new ArrayList<>();
 
-        List<Map<String,Object> > ret = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
+        List<Map<String, Object>> ret = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
 
-        for(Map<String,Object> m : ret) {
-            list.add((String)m.get(EAV_A_USER_CLASS.META_NAME.getName()));
-        }
+        for (Map<String, Object> m : ret) list.add((String) m.get(EAV_A_USER_CLASS.META_NAME.getName()));
 
         return list;
     }
 
     @Override
-    public List<Long> getAllowedRefs(long portalUserId, String meta){
+    public List<Long> getAllowedRefs(long portalUserId, String meta) {
         Select select = context.select(EAV_A_USER_REF.ENTITY_ID)
                 .from(EAV_A_USER_REF)
                 .where(EAV_A_USER_REF.USER_ID.eq(portalUserId))
                 .and(EAV_A_USER_REF.META_NAME.eq(meta));
 
-        List<Long> list = new ArrayList<Long>();
+        List<Long> list = new ArrayList<>();
 
-        List<Map<String,Object> > ret = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
+        List<Map<String, Object>> ret = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
 
-        for(Map<String,Object> m : ret) {
-            list.add( ((BigDecimal)m.get(EAV_A_USER_REF.ENTITY_ID.getName())).longValue());
-        }
+        for (Map<String, Object> m : ret)
+            list.add(((BigDecimal) m.get(EAV_A_USER_REF.ENTITY_ID.getName())).longValue());
 
         return list;
     }
 
     @Override
     public PortalUser getUser(long userId) {
-        Select select = context.select()
-                .from(EAV_A_USER)
-                .where(EAV_A_USER.USER_ID.eq(userId));
+        Select select = context.select().from(EAV_A_USER).where(EAV_A_USER.USER_ID.eq(userId));
 
         List<Map<String, Object>> list = queryForListWithStats(select.getSQL(), select.getBindValues().toArray());
 
-        if(list == null || list.isEmpty())
+        if (list == null || list.isEmpty())
             return null;
 
         Map<String, Object> row = list.get(0);
@@ -399,6 +309,7 @@ public class UserDaoImpl extends JDBCSupport implements IUserDao
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public List<PortalUser> getPortalUsersHavingAccessToCreditor(Creditor creditor) {
         Table root = EAV_A_USER.as("root");
         Table creditors = EAV_A_CREDITOR_USER.as("creditors");
